service-b/requirements.txt
fastapi
uvicorn
python-dotenv
neo4j
requests